UI
Worked on UI. Madea thing to plot from / to. Holy hell was that hard, however I think I finally have it working for inclanations, Also something weird happend with check / get ID, where for the 1960 TLE it was all holding text. I have found a bettwe way to remove rows in arrays so i will try to implement that method on the get check next